tests/wpt/tests/html/canvas/offscreen/text/2d.text.lang.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.lang.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.lang.valid</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.lang = '';
  _assertSame(ctx.lang, '', "ctx.lang", "''");

  ctx.lang = 'inherit';
  _assertSame(ctx.lang, 'inherit', "ctx.lang", "'inherit'");

  ctx.lang = 'en-US';
  _assertSame(ctx.lang, 'en-US', "ctx.lang", "'en-US'");

  ctx.lang = 'not-a-real-lang';
  _assertSame(ctx.lang, 'not-a-real-lang', "ctx.lang", "'not-a-real-lang'");
  t.done();

});
</script>
